Feat: FileStorage 관련 예외 처리 수정 및 로깅 추가
- GlobalExceptionHandler에서 File 관련 예외처리 FileExceptionHandler로 분리
- ResultCode 추가

+++ src/main/java/com/sprint/mission/discodeit/storage/LocalBinaryContentStorage.java
@@ -1,7 +1,12 @@
 package com.sprint.mission.discodeit.storage;
 
 import com.sprint.mission.discodeit.dto.service.binarycontent.FindBinaryContentResult;
-import com.sprint.mission.discodeit.exception.RestExceptions;
+import com.sprint.mission.discodeit.exception.file.DuplicateFilePathException;
+import com.sprint.mission.discodeit.exception.file.FileDeleteException;
+import com.sprint.mission.discodeit.exception.file.FileNotFoundException;
+import com.sprint.mission.discodeit.exception.file.FileReadException;
+import com.sprint.mission.discodeit.exception.file.FileWriterException;
+import com.sprint.mission.discodeit.exception.file.InitDirectoryException;
 import jakarta.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
@@ -9,6 +14,7 @@ import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Map;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,7 +22,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.InvalidMediaTypeException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
@@ -35,7 +40,8 @@ public class LocalBinaryContentStorage implements BinaryContentStorage {
     try {
       Files.createDirectories(root);
     } catch (IOException e) {
-      throw RestExceptions.INIT_DIRECTORY_ERROR;
+      log.error("Directory init failed: (root: {})", root);
+      throw new InitDirectoryException(Map.of("root", root));
     }
   }
 
@@ -43,23 +49,33 @@ public class LocalBinaryContentStorage implements BinaryContentStorage {
   @Override
   public UUID put(UUID binaryContentId, byte[] bytes) {
     Path filePath = resolvePath(binaryContentId);
+    if (Files.exists(filePath)) {
+      log.warn("File write failed: duplicate file path (filepath: {})", filePath);
+      throw new DuplicateFilePathException(Map.of("filePath", filePath));
+    }
     try {
       Files.write(filePath, bytes);
       return binaryContentId;
     } catch (IOException e) {
-      log.error("파일 저장 실패: {}", binaryContentId, e);
-      throw RestExceptions.FILE_WRITE_ERROR;
+      log.error("File write failed (binaryContentId: {}, path: {}, size: {})", binaryContentId,
+          filePath, bytes.length);
+      throw new FileWriterException(
+          Map.of("binaryContentId", binaryContentId, "path", filePath, "size", bytes.length));
     }
   }
 
   @Override
   public InputStream get(UUID binaryContentId) {
     Path filePath = resolvePath(binaryContentId);
+    if (Files.notExists(filePath)) {
+      log.warn("File read failed: file not found (filepath: {})", filePath);
+      throw new FileNotFoundException(Map.of("filePath", filePath));
+    }
     try {
       return Files.newInputStream(filePath);
     } catch (IOException e) {
-      log.error("파일 읽기 실패: {}", binaryContentId, e);
-      throw RestExceptions.FILE_READ_ERROR;
+      log.error("File read failed (binaryContentId: {}, path: {})", binaryContentId, filePath);
+      throw new FileReadException(Map.of("binaryContentId", binaryContentId, "path", filePath));
     }
   }
 
@@ -69,34 +85,30 @@ public class LocalBinaryContentStorage implements BinaryContentStorage {
     try {
       Files.deleteIfExists(filePath);
     } catch (IOException e) {
-      log.error("파일 삭제 실패: {}", binaryContentId, e);
-      throw RestExceptions.FILE_DELETE_ERROR;
+      log.error("File delete failed (binaryContentId: {}, path: {})", binaryContentId, filePath);
+      throw new FileDeleteException(Map.of("binaryContentId", binaryContentId, "path", filePath));
     }
   }
 
   @Override
   public ResponseEntity<Resource> download(FindBinaryContentResult findBinaryContentResult) {
-    try {
-      // 대용량 파일을 한번에 넘기면 부하가 크므로, inputStream에 담음
-      InputStream inputStream = get(findBinaryContentResult.id());
-      // InputStream을 Resource로 감싸서 넘기면, 스프링이 자동으로 스트리밍 다운로드 처리를 해줌 (다운로드 진행바 생각)
-      // 대용량 파일을 조각조각 읽기(InputStream) → 브라우저에 실시간 전송(Resource로 감싸서 Spring이 스트리밍 처리)
-      Resource resource = new InputStreamResource(inputStream);
+    // 대용량 파일을 한번에 넘기면 부하가 크므로, inputStream에 담음
+    InputStream inputStream = get(findBinaryContentResult.id());
+    // InputStream을 Resource로 감싸서 넘기면, 스프링이 자동으로 스트리밍 다운로드 처리를 해줌 (다운로드 진행바 생각)
+    // 대용량 파일을 조각조각 읽기(InputStream) → 브라우저에 실시간 전송(Resource로 감싸서 Spring이 스트리밍 처리)
+    Resource resource = new InputStreamResource(inputStream);
 
-      String encodedFilename = URLEncoder.encode(findBinaryContentResult.filename(),
-              StandardCharsets.UTF_8)
-          .replaceAll("\\+", "%20"); // 공백 처리
+    String encodedFilename = URLEncoder.encode(findBinaryContentResult.filename(),
+            StandardCharsets.UTF_8)
+        .replaceAll("\\+", "%20"); // 공백 처리
 
-      return ResponseEntity.ok()
-          .header(HttpHeaders.CONTENT_DISPOSITION,
-              "inline; filename=\"" + encodedFilename + "\"")
-          // 프론트단에서 BinaryContent download API를 날려 사진을 가져와 바로 렌더링하는 것으로 보임 -> 이를 위해 inline으로 지정
-          .contentType(MediaType.parseMediaType(findBinaryContentResult.contentType()))
-          .header(HttpHeaders.CONTENT_LENGTH, String.valueOf(findBinaryContentResult.size()))
-          .body(resource);
-    } catch (InvalidMediaTypeException e) {
-      throw RestExceptions.UNSUPPORTED_MEDIA_TYPE;
-    }
+    return ResponseEntity.ok()
+        .header(HttpHeaders.CONTENT_DISPOSITION,
+            "inline; filename=\"" + encodedFilename + "\"")
+        // 프론트단에서 BinaryContent download API를 날려 사진을 가져와 바로 렌더링하는 것으로 보임 -> 이를 위해 inline으로 지정
+        .contentType(MediaType.parseMediaType(findBinaryContentResult.contentType()))
+        .header(HttpHeaders.CONTENT_LENGTH, String.valueOf(findBinaryContentResult.size()))
+        .body(resource);
   }
 
   private Path resolvePath(UUID id) {
